requirements.txt

gymnasium
stable-baselines3
pygame
pyopengl
numpy
imageio
tensorflow